scansettlr-ocr/requirements.txt
fastapi
uvicorn
pytesseract
Pillow
opencv-python-headless
python-multipart